bugfix: keep MateGated params under musa namespace.
Own MUSA-only GDN params in gdn_ops.h so shared param.h stays untouched,
and route layer call sites through kernel::musa:: with consistent A_log fields.
