+++ xllm/core/kernels/musa/ops_api.h
@@ -17,10 +17,7 @@ limitations under the License.
 
 #include <torch/torch.h>
 
-#include <cstdint>
 #include <optional>
-#include <tuple>
-#include <vector>
 
 namespace xllm::kernel::musa {
 
@@ -41,48 +38,8 @@ struct Fp8BlockMatmulParams {
   std::optional<torch::Tensor> output;
 };
 
-struct MateGatedDeltaRulePrefillParams {
-  torch::Tensor q;
-  torch::Tensor k;
-  torch::Tensor v;
-  torch::Tensor g;
-  torch::Tensor beta;
-  std::optional<float> scale = std::nullopt;
-  std::optional<torch::Tensor> initial_state = std::nullopt;
-  std::optional<torch::Tensor> cu_seqlens = std::nullopt;
-  std::optional<std::vector<int32_t>> cu_seqlens_host = std::nullopt;
-  std::optional<torch::Tensor> output = std::nullopt;
-  std::optional<torch::Tensor> final_state = std::nullopt;
-  std::optional<torch::Tensor> kkt_output = std::nullopt;
-  bool use_qk_l2norm_in_kernel = true;
-  bool allow_inplace_qk_l2norm = false;
-};
-
-struct MateGatedDeltaRuleDecodeParams {
-  torch::Tensor mixed_qkv;
-  torch::Tensor state;
-  torch::Tensor a_log;
-  torch::Tensor a;
-  torch::Tensor dt_bias;
-  torch::Tensor b;
-  torch::Tensor state_indices;
-  int64_t num_k_heads = 0;
-  int64_t num_v_heads = 0;
-  int64_t head_k_dim = 0;
-  int64_t head_v_dim = 0;
-  double scale = 0.0;
-  bool use_qk_l2norm = true;
-  std::optional<torch::Tensor> decode_output = std::nullopt;
-};
-
 torch::Tensor matmul(MatmulParams& params);
 
 torch::Tensor fp8_block_matmul(Fp8BlockMatmulParams& params);
-
-std::tuple<torch::Tensor, torch::Tensor> per_token_group_quant_fp8(
-    const torch::Tensor& input,
-    int64_t group_size);
-
-void mul_sigmoid_gate_inplace(torch::Tensor& out, const torch::Tensor& gate);
 
 }  // namespace xllm::kernel::musa
